Making the BinaryTree a gem
